+++ src/animations.js
@@ -7,7 +7,12 @@ import {
 
 import {
   fadeTransition,
-  appearUpTransition
+  appearUpTransition,
+  appearDownTransition,
+  appearLeftTransition,
+  appearRightTransition,
+  appearInTransition,
+  appearOutTransition
 } from './transitionDefinitions'
 
 export const Fade = ({
@@ -49,6 +54,121 @@ export const AppearUp = ({
   </Animate>
 )
 AppearUp.defaultProps = {
+  timeout: pageDefaultTiming,
+  defaultStyle: {},
+  distance: defaultSlideDistance
+}
+
+export const AppearDown = ({
+  in: inProp,
+  timeout,
+  defaultStyle,
+  distance,
+  children,
+  ...props
+}) => (
+  <Animate
+    in={inProp}
+    timeout={timeout}
+    transitionStyles={appearDownTransition({timeout, defaultStyle, distance})}
+    {...props}
+  >
+    {children}
+  </Animate>
+)
+AppearDown.defaultProps = {
+  timeout: pageDefaultTiming,
+  defaultStyle: {},
+  distance: defaultSlideDistance
+}
+
+export const AppearLeft = ({
+  in: inProp,
+  timeout,
+  defaultStyle,
+  distance,
+  children,
+  ...props
+}) => (
+  <Animate
+    in={inProp}
+    timeout={timeout}
+    transitionStyles={appearLeftTransition({timeout, defaultStyle, distance})}
+    {...props}
+  >
+    {children}
+  </Animate>
+)
+AppearLeft.defaultProps = {
+  timeout: pageDefaultTiming,
+  defaultStyle: {},
+  distance: defaultSlideDistance
+}
+
+export const AppearRight = ({
+  in: inProp,
+  timeout,
+  defaultStyle,
+  distance,
+  children,
+  ...props
+}) => (
+  <Animate
+    in={inProp}
+    timeout={timeout}
+    transitionStyles={appearRightTransition({timeout, defaultStyle, distance})}
+    {...props}
+  >
+    {children}
+  </Animate>
+)
+AppearRight.defaultProps = {
+  timeout: pageDefaultTiming,
+  defaultStyle: {},
+  distance: defaultSlideDistance
+}
+
+export const AppearIn = ({
+  in: inProp,
+  timeout,
+  defaultStyle,
+  distance,
+  children,
+  ...props
+}) => (
+  <Animate
+    in={inProp}
+    timeout={timeout}
+    transitionStyles={appearInTransition({timeout, defaultStyle, distance})}
+    {...props}
+  >
+    {children}
+  </Animate>
+)
+AppearIn.defaultProps = {
+  timeout: pageDefaultTiming,
+  defaultStyle: {},
+  distance: defaultSlideDistance
+}
+
+export const AppearOut = ({
+  in: inProp,
+  timeout,
+  defaultStyle,
+  distance,
+  children,
+  ...props
+}) => (
+  <Animate
+    in={inProp}
+    timeout={timeout}
+    transitionStyles={appearOutTransition({timeout, defaultStyle, distance})}
+    {...props}
+  >
+    {children}
+  </Animate>
+)
+AppearOut.defaultProps = {
   timeout: pageDefaultTiming,
   defaultStyle: {},
   distance: defaultSlideDistance