+++ changelog.md
@@ -2,7 +2,7 @@
 #
 
 * _add_ expose style helpers
-* _add_ add appear up component
+* _add_ add appear components
 * _add_ fade animation component
 
 # 0.1.0